Move tensor operations to other file
Move tensor operations to other file and reduce
the inference steps of z image denoiser test.
